# server-chat/docs/swagger.yaml
basePath: /
definitions:
...
  apidocs.AuthCredentials:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: longpassword123
        type: string
    type: object
  apidocs.AuthSuccess:
    properties:
      data:
        properties:
          access_token:
            type: string
          refresh_token:
            type: string
          user:
            $ref: '#/definitions/apidocs.UserPayload'
        type: object
      success:
        type: boolean
    type: object
...
  apidocs.ErrorEnvelope:
    properties:
      error:
        properties:
          code:
            example: INVALID_BODY
            type: string
          message:
            example: Invalid or malformed request body
            type: string
        type: object
      success:
        example: false
        type: boolean
    type: object
...
  apidocs.OKMessage:
    properties:
      data:
        properties:
          message:
            example: Done
            type: string
        type: object
      success:
        example: true
        type: boolean
    type: object
...
  apidocs.RefreshRequest:
    properties:
      refresh_token:
        example: a1b2c3...
        type: string
    type: object
  apidocs.RefreshSuccess:
    properties:
      data:
        properties:
          access_token:
            type: string
          refresh_token:
            type: string
        type: object
      success:
        type: boolean
    type: object
  apidocs.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: longpassword123
        type: string
      username:
        example: john_doe
        type: string
    type: object
...
  apidocs.UserPayload:
    properties:
      avatar:
        example: https://...
        type: string
      bio:
        example: Hello!
        type: string
      email:
        example: user@example.com
        type: string
      id:
        type: integer
      is_online:
        type: boolean
      last_seen_at:
        type: string
      username:
        example: john_doe
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST + WebSocket backend for realtime chat (rooms, JWT auth, message
    history, reactions, notifications).
  title: Chat Server API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.AuthCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.AuthSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorEnvelope'
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Specific refresh_token to revoke; omit to revoke all
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.OKMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Logout (revoke refresh token)
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.RefreshSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorEnvelope'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials + optional username
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidocs.AuthSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apidocs.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorEnvelope'
      summary: Register account
      tags:
      - auth
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: Bearer followed by a space and the token
      value.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"